fast_api/requirements.txt
PyJWT==2.4.0
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.1
gunicorn==20.1.0
httptools==0.4.0
orjson==3.4.1
pydantic==1.9.0
python-dotenv==0.20.0
requests==2.25.0
uvicorn==0.12.2
uvloop==0.14.0